requirements.txt
discord.py
flask
aiohttp
python-dotenv
beautifulsoup4
pillow
pytz